fix: Fixed wrong typing

+++ db/entities/handledMessage.go
@@ -5,6 +5,6 @@ import "gorm.io/gorm"
 type HandledMessage struct {
 	gorm.Model
 	OriginalMessageId int
-	HandledAt         int
+	HandledAt         int64
 	MarshaledMessage  string
 }
